requirements/base.txt
django==4.2.*
python-decouple==3.8
openai
